Remove redunant code
Removed product filter, as the backend handles this
Removed checkUnput function and related event listener, as the search button should always be enabled to allow the user to reset the page with no search query. Go to amazon and search something, delete your search and hit the search button. It does exactly this.